+++ native_c/intel/packet/gcm_siv_pc/gcm_siv_pc.h
@@ -13,7 +13,7 @@
 #define BC_FIPS_C_GCM_SIV_PC_H
 
 packet_err* gcm_siv_pc_process_packet(bool encryption, uint8_t *key, size_t keysize, uint8_t *iv,
-                                  uint8_t *aad, size_t aadLen, uint8_t *p_in, size_t inLen, uint8_t *p_out, size_t *outputLen);
+                                  uint8_t *aad, int aadLen, uint8_t *p_in, size_t inLen, uint8_t *p_out, size_t *outputLen);
